clean up and factor out code
